move the mock action handler to the base test to share between a few different areas
