+++ tests/sentry/workflow_engine/test_base.py
@@ -25,7 +25,7 @@ from sentry.workflow_engine.models import (
 )
 from sentry.workflow_engine.models.data_condition import Condition
 from sentry.workflow_engine.registry import data_source_type_registry
-from sentry.workflow_engine.types import DetectorPriorityLevel
+from sentry.workflow_engine.types import ActionHandler, DetectorPriorityLevel
 from tests.sentry.issues.test_utils import OccurrenceTestMixin
 
 try:
@@ -41,6 +41,34 @@ class MockModel(Model):
 
     class Meta:
         app_label = "fixtures"
+
+
+class MockActionHandler(ActionHandler):
+    config_schema = {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "description": "The configuration schema for a Action Configuration",
+        "type": "object",
+        "properties": {
+            "foo": {
+                "type": ["string"],
+            },
+        },
+        "required": ["foo"],
+        "additionalProperties": False,
+    }
+
+    data_schema = {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "description": "The configuration schema for a Action Data",
+        "type": "object",
+        "properties": {
+            "baz": {
+                "type": ["string"],
+            },
+        },
+        "required": ["baz"],
+        "additionalProperties": False,
+    }
 
 
 class BaseWorkflowTest(TestCase, OccurrenceTestMixin):
